# api/openapi.yaml
openapi: 3.1.0
info:
  title: Shubayr API
  version: 11.0.0
  description: >
    REST/JSON API contract for Shubayr — an online multi-section store
    (single-owner, white-label). This document is the SHARED CONTRACT between
    the backend (NestJS) and the clients (Next.js web, Flutter mobile/admin).
    It must stay in sync with infra/db/schema.sql. Changing an endpoint's shape
    is a reviewed PR that updates this file first.

    Access model v2 (breaking): every bearer token has `surface=admin|app`.
    Admin tokens come only from username/password login and staff operations
    require both `surface=admin` and the route's `x-permission`. App tokens come
    only from phone OTP, contain the server-resolved single app role, and never
    authorize admin operations. App clients must not send a role during OTP.
    Work roles (`delivery_agent`, `order_monitor`) cannot use purchase functions.

    Validation and errors: malformed or invalid request data returns HTTP 422
    with code `VALIDATION_FAILED` and field-level errors. All non-2xx JSON
    responses use the Error schema. PATCH distinguishes omission from null:
    omitted fields are unchanged, while null is accepted only for explicitly
    nullable fields and clears that field.

    Money policy: every money-bearing resource identifies its ISO currency.
    Amounts are stored exactly and displayed at that currency's configured
    precision (IQD 0, USD 2). The ledger preserves original-currency amount and
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List the current user's orders
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/OrderStatus" } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderPage" }
    post:
      tags: [Orders]
      summary: Place a Cash-on-Delivery order
      description: >
        Atomically reprices and consumes the caller's server cart using server-time
        effective prices, creates status=pending, a pending COD payment and a
        minimal delivery record. Sellable stock is reduced by a product/variant
        hold; FEFO batch assignment and picking are reserved for the inventory
        slice. A repeated Idempotency-Key with the same checkout request returns
        the original order, even after the cart is cleared; a different request
        with that key returns 409. Without a key, retries can create a new order.
        The selected owned address and contact_phone are copied into immutable
        order snapshot fields in the same transaction. Later address edits do
        not alter snapshots; deletion clears address_id but retains snapshots.
      parameters:
        - name: Idempotency-Key
          in: header
          description: Optional printable ASCII key, 1-128 characters, unique per customer.
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id]
              properties:
                address_id: { type: string, format: uuid }
                coupon_code: { type: [string, "null"] }
                payment_method: { type: string, enum: [cod], default: cod }
                accepted_price_versions:
                  type: array
                  description: Explicit acceptance tokens returned by a prior PRICE_CHANGED conflict.
                  items:
                    type: object
                    additionalProperties: false
                    required: [variant_id, price_version]
                    properties:
                      variant_id: { type: string, format: uuid }
                      price_version: { type: string, minLength: 1, maxLength: 64 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Cart empty, item unavailable, coupon invalid, or idempotency conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Order detail
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/{id}/track:
    get:
      tags: [Orders]
      summary: Order status timeline
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      responses:
        "200":
          description: Ordered list of status-change events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderTracking" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel a pending or confirmed order (releases its simple stock holds)
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version]
              properties:
                version: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Deprecated alias for the staff pre-dispatch transition endpoint
      deprecated: true
      x-surface: admin
      x-permission-by-status:
        confirmed: orders.accept
        preparing: orders.prepare
        ready_for_dispatch: orders.mark_ready
        dispatched: orders.handover
      description: >
        Supports the same pre-dispatch transitions as
        PATCH /admin/orders/{id}/status. Use the dedicated admin endpoint for
        new clients.
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, version]
              properties:
                status: { type: string, enum: [confirmed, preparing, ready_for_dispatch, dispatched] }
                version: { type: integer, minimum: 1 }
                note: { type: [string, "null"], maxLength: 500 }
                below_cost_override_reason: { type: [string, "null"], maxLength: 500 }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOrder" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /orders/{id}/cancellation-request:
    post:
      tags: [Orders]
      summary: Request cancellation after the pending state
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version, reason]
              properties:
                version: { type: integer, minimum: 1 }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: Order with pending cancellation request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /orders/{id}/shortage-response:
    post:
      tags: [Orders]
      summary: Accept or deny a proposed preparation quantity reduction
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version, decision]
              properties:
                version: { type: integer, minimum: 1 }
                decision: { type: string, enum: [accepted, denied] }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }

  # ---------------- REVIEWS ----------------
...
  /monitor/orders/{id}:
    get:
      tags: [Monitoring]
      summary: Read-only monitor order detail without product images
      x-surface: app
      x-app-role: order_monitor
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      responses:
        "200":
          description: Monitor-safe order detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonitorOrderDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }

  # ---------------- ADMIN: REPORTS & SETTINGS ----------------
...
  /admin/orders/{id}:
    get:
      tags: [Admin]
      summary: Get a full order-management detail
      x-permission: orders.view
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      responses:
        "200":
          description: Order with immutable lines, customer, shipping snapshot, payment, delivery, agent, and timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOrder" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Advance an order through its pre-dispatch lifecycle
      x-surface: admin
      x-permission-by-status:
        confirmed: orders.accept
        preparing: orders.prepare
        ready_for_dispatch: orders.mark_ready
        dispatched: orders.handover
      description: >
        Legal path: pending to confirmed to preparing to ready_for_dispatch to
        dispatched. Dispatch requires the existing current delivery to have an
        agent assigned through PATCH /deliveries/{id}/assign; it moves that same
        delivery to out_for_delivery and converts checkout stock holds from held
        to deducted. Delivery agent transitions own delivered, failed and returned.
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, version]
              properties:
                status:
                  type: string
                  enum: [confirmed, preparing, ready_for_dispatch, dispatched]
                version: { type: integer, minimum: 1 }
                note: { type: [string, "null"], maxLength: 500 }
                below_cost_override_reason: { type: [string, "null"], maxLength: 500 }
      responses:
        "200":
          description: Updated order detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOrder" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/PostingConflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/orders/{id}/cancel:
    post:
      tags: [Admin]
      summary: Cancel a pre-dispatch order and release its stock holds
      x-permission: orders.cancel
      description: >
        Allowed only from pending, confirmed, preparing, or ready_for_dispatch.
        A non-empty reason is stored in the timeline and audit trail. Dispatched,
        delivered, failed, returned, and paid COD orders cannot be cancelled.
      parameters: [ { $ref: "#/components/parameters/PathId" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version, reason]
              properties:
                version: { type: integer, minimum: 1 }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: Cancelled order detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOrder" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
...
    NotFound: { description: Not found (`NOT_FOUND`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Conflict: { description: State conflict (`CONFLICT`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    PostingConflict:
      description: State conflict (`CONFLICT`) or a closed accounting period (`PERIOD_CLOSED`)
      content:
        application/json:
          schema:
            oneOf:
              - { $ref: "#/components/schemas/Error" }
              - { $ref: "#/components/schemas/PeriodClosedError" }
    Validation: { description: "Invalid request data (`VALIDATION_FAILED`, HTTP 422)", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Forbidden: { description: Missing permission (`FORBIDDEN`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [status, code, message, errors]
      properties:
        status: { type: integer, example: 422 }
        code: { type: string, example: VALIDATION_FAILED }
        message: { type: string }
        errors:
          type: array
          description: Empty for non-validation errors.
          items: { $ref: "#/components/schemas/FieldError" }
    PeriodClosedError:
      type: object
      additionalProperties: false
      required: [status, code, message, errors, period]
      properties:
        status: { type: integer, enum: [409] }
        code: { type: string, enum: [PERIOD_CLOSED] }
        message: { type: string }
        errors: { type: array, maxItems: 0, items: { $ref: "#/components/schemas/FieldError" } }
        period: { type: string, pattern: '^\d{4}-\d{2}$', example: "2026-11" }
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string, example: items.0.quantity }
        code: { type: string, example: min }
        message: { type: string, example: quantity must not be less than 1 }
        requested: { type: string, description: Exact requested quantity for an inventory conflict. }
        available: { type: string, description: Exact currently allocatable quantity for an inventory conflict. }
        variant_id: { type: string, format: uuid, description: SKU involved in an inventory conflict. }
        product_id: { type: string, format: uuid }
        sku: { type: string }
        old_price: { type: number }
        new_price: { type: number }
        old_price_version: { type: string }
        new_price_version: { type: string }
        current_status: { $ref: "#/components/schemas/OrderStatus" }
        current_version: { type: integer, minimum: 1 }
        price: { type: number }
        threshold_percent: { type: number }
        cost: { type: number }
        minimum_price: { type: number }
    Money:
      type: number
      description: Decimal money amount displayed at the associated currency precision (IQD 0, USD 2).
...
    Pagination:
      type: object
      description: Reusable pagination envelope. List endpoints add a `data` array of the resource.
      required: [page, per_page, total]
      properties:
        page: { type: integer, example: 1 }
        per_page: { type: integer, example: 20 }
        total: { type: integer, example: 0 }
...
    OrderStatus:
      type: string
      description: Canonical order lifecycle status (matches orders.status in schema.sql).
      enum:
        - pending
        - confirmed
        - preparing
        - ready_for_dispatch
        - dispatched
        - delivered
        - failed
        - rejected
        - cancelled
        - return_requested
        - returned
    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid, description: "order_item_id — reference this from reviews & returns" }
        product_id: { type: string, format: uuid }
        variant_id: { type: string, format: uuid }
        product_name_ar: { type: string, description: "Arabic product name captured when the order is created." }
        product_name_en: { type: string, description: "English product name captured when the order is created." }
        image_url: { type: [string, "null"], format: uri, description: "Primary product image URL captured when the order is created." }
        quantity: { type: number, multipleOf: 0.001 }
        unit_price: { $ref: "#/components/schemas/Money", description: "The product's effective_price captured when the order is created, so a later discount edit or window expiry never rewrites a placed order." }
        price_version: { type: string, maxLength: 64 }
        line_total: { $ref: "#/components/schemas/Money" }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        reviewed: { type: boolean, readOnly: true, description: True when the caller already has a product review for this order item. }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_number: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        version: { type: integer, minimum: 1, description: Required on every state-changing order request. }
        payment_method: { type: string, enum: [cod] }
        address_id: { type: [string, "null"], format: uuid, description: "Original address reference; null if that address was later deleted. Delivery snapshots remain unchanged." }
        delivery_id: { type: [string, "null"], format: uuid, description: Current delivery reachable by this order's owner. }
        subtotal: { $ref: "#/components/schemas/Money" }
        delivery_fee: { $ref: "#/components/schemas/Money" }
        discount: { $ref: "#/components/schemas/Money" }
        total: { $ref: "#/components/schemas/Money" }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        delivery_contact_phone: { type: string, description: Immutable checkout snapshot. }
        delivery_address_label: { type: [string, "null"], description: Immutable checkout snapshot. }
        delivery_city: { type: string, description: Immutable checkout snapshot. }
        delivery_area: { type: [string, "null"], description: Immutable checkout snapshot. }
        delivery_street: { type: [string, "null"], description: Immutable checkout snapshot. }
        delivery_details: { type: [string, "null"], description: Immutable checkout snapshot. }
        delivery_lat: { type: [number, "null"], format: double, description: Immutable checkout snapshot. }
        delivery_lng: { type: [number, "null"], format: double, description: Immutable checkout snapshot. }
        placed_at: { type: string, format: date-time }
        acceptance_deadline: { type: [string, "null"], format: date-time }
        auto_cancel_deadline: { type: [string, "null"], format: date-time }
        late_for_acceptance: { type: boolean }
        cancellation_request:
          oneOf:
            - type: "null"
            - type: object
              properties:
                status: { type: string, enum: [pending, approved, denied] }
                reason: { type: [string, "null"] }
                requested_at: { type: [string, "null"], format: date-time }
                resolved_at: { type: [string, "null"], format: date-time }
                resolution_note: { type: [string, "null"] }
        price_change_info: { type: [object, "null"], additionalProperties: true }
        inventory_attention_required: { type: boolean }
        attention_details: { type: [object, "null"], additionalProperties: true }
        timeline:
          type: array
          items:
            type: object
            properties:
              status: { $ref: "#/components/schemas/OrderStatus" }
              note: { type: [string, "null"] }
              at: { type: string, format: date-time }
        retrievals:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              document_number: { type: string }
              status: { type: string, enum: [open, partially_received, received, closed] }
              outcome: { type: string, enum: [cancel, retry] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
    OrderPage:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Order" }
    AdminOrder:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [customer, shipping_snapshot, payments, delivery, status_events]
          properties:
            customer:
              type: object
              required: [id, name, phone, email]
              properties:
                id: { type: string, format: uuid }
                name: { type: [string, "null"] }
                phone: { type: string }
                email: { type: [string, "null"] }
            shipping_snapshot:
              type: object
              description: Immutable checkout shipping and contact values.
              properties:
                contact_phone: { type: string }
                address_label: { type: [string, "null"] }
                city: { type: string }
                area: { type: [string, "null"] }
                street: { type: [string, "null"] }
                details: { type: [string, "null"] }
                lat: { type: [number, "null"] }
                lng: { type: [number, "null"] }
            payments:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  method: { type: string, enum: [cod] }
                  status: { type: string, enum: [pending, paid, refunded, failed] }
                  amount: { $ref: "#/components/schemas/Money" }
                  currency: { type: string, pattern: '^[A-Z]{3}$' }
                  paid_at: { type: [string, "null"], format: date-time }
            delivery:
              type: [object, "null"]
              properties:
                id: { type: string, format: uuid }
                status: { type: string, enum: [assigned, out_for_delivery, delivered, failed, returned] }
                delivery_fee: { $ref: "#/components/schemas/Money" }
                currency: { type: string, pattern: '^[A-Z]{3}$' }
                dispatched_at: { type: [string, "null"], format: date-time }
                delivered_at: { type: [string, "null"], format: date-time }
                failure_reason: { type: [string, "null"], maxLength: 500 }
                failed_at: { type: [string, "null"], format: date-time }
                retry_count: { type: integer, minimum: 0 }
                agent:
                  type: [object, "null"]
                  properties:
                    id: { type: string, format: uuid }
                    name: { type: [string, "null"] }
                    phone: { type: string }
                    email: { type: [string, "null"] }
            status_events:
              type: array
              items:
                type: object
                properties:
                  status: { $ref: "#/components/schemas/OrderStatus" }
                  note: { type: [string, "null"] }
                  at: { type: string, format: date-time }
...
    MonitorOrderDetail:
      type: object
      description: Intentionally excludes product image fields.
      required: [id, order_number, status, customer, shipping_snapshot, items, subtotal, delivery_fee, discount, total, currency, payment_method, placed_at]
      properties:
        id: { type: string, format: uuid }
        order_number: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        customer:
          type: object
          required: [name, phone]
          properties:
            name: { type: [string, "null"] }
            phone: { type: string }
        shipping_snapshot:
          type: object
          properties:
            contact_phone: { type: string }
            address_label: { type: [string, "null"] }
            city: { type: string }
            area: { type: [string, "null"] }
            street: { type: [string, "null"] }
            details: { type: [string, "null"] }
            lat: { type: [number, "null"] }
            lng: { type: [number, "null"] }
        items:
          type: array
          items:
            type: object
            required: [id, product_id, variant_id, product_name_ar, product_name_en, quantity, unit_price, line_total, currency]
            properties:
              id: { type: string, format: uuid }
              product_id: { type: string, format: uuid }
              variant_id: { type: [string, "null"], format: uuid }
              product_name_ar: { type: string }
              product_name_en: { type: string }
              quantity: { type: number, multipleOf: 0.001 }
              unit_price: { $ref: "#/components/schemas/Money" }
              line_total: { $ref: "#/components/schemas/Money" }
              currency: { type: string, pattern: '^[A-Z]{3}$' }
        subtotal: { $ref: "#/components/schemas/Money" }
        delivery_fee: { $ref: "#/components/schemas/Money" }
        discount: { $ref: "#/components/schemas/Money" }
        total: { $ref: "#/components/schemas/Money" }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        payment_method: { type: string, enum: [cod] }
        placed_at: { type: string, format: date-time }
    OrderTracking:
      type: object
      properties:
        order_id: { type: string, format: uuid }
        events:
          type: array
          items:
            type: object
            properties:
              status: { $ref: "#/components/schemas/OrderStatus" }
              note: { type: [string, "null"] }
              at: { type: string, format: date-time }
...